requirements.txt
streamlit==1.24.1
numpy==1.25.1
pandas==2.0.3
Pillow==9.5.0